requirements.txt
certifi==2025.1.31
cftime==1.6.4.post1
DateTime==5.5
filelock==3.17.0
fsspec==2025.3.0
jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
netCDF4==1.7.2
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
pillow==11.1.0
python-sofa==0.2.0
pytz==2025.1
scipy==1.13.1
sympy==1.13.1
torch==2.6.0
torch-summary==1.4.5
torchaudio==2.6.0
torchvision==0.21.0
triton==3.2.0
typing-extensions==4.12.2
zope.interface==7.2
